Add files via upload
You can select widescreen on/off in the Options menu. CRT ONLY - HD connections on widescreen monitors will not notice the change.

## New Features are in line with JAGDOOMEX minus the gamedrive testing options.  may add later.

You can select widescreen on/off in the Options menu. This makes Doom look the
proper aspect ratio on 16:9 TVs. Off assumes a standard 4:3 TV. Defaults to off,
and is saved in the e2prom.

Chilly has fixed the Arms indicators so that they show the proper keypad number. At
the same time, He's added colors to indicate some status info. The number is
GREEN if the weapon is equipped, BLUE if the required ammo is more than half
the max allowed count (with or without the backpack), YELLOW if the required
ammo is less than half the max allowed count, and RED if the weapon is out of
ammo. At a glance, you get vital information. He can make this an option if
people prefer not having colored indicators. Let Chilly or Tolbat know how you feel about
this.

## New Controls

C/B/A/PAUSE/OPTION remain as set in the game. The differences arise in the
number pad, which also affects the ProPad controller.
```
1 = select PISTOL
2 = FORWARD
3 = select SHOTGUN
4 = STRAFE LEFT
5 = BACKWARD
6 = STRAFE RIGHT
7 = select CHAINGUN
8 = select MISSILE LAUNCHER
9 = select PLASMA RIFLE
* = toggle between FIST/CHAINSAW
0 = select BFG9000
# = select AUTOMAP
* + # = reset game
```
4 and 6 are the same as the LEFT and RIGHT shoulder buttons on the ProPad, which
is why they are set for strafing. I set 2 and 5 to forward and backward because
Chilly will be adding mouse support, and having those four buttons arranged as they
are will make it easier to play like using a keyboard and mouse on the PC. The
buttons 7, 8, and 9 map to the ProPad buttons Z, Y, and X, respectively. This
gives you quick access to the heavier weapons in Doom. Decided not to put the
BFG in there as it usually has limited ammo, so you really don't want to be 
accidentally selecting it. Likewise, selecting the fist or chainsaw is fairly
rare, so it could be one of the out-of-the-way buttons. When Doom 2 support gets
added, the SHOTGUN button will toggle between the shotgun and the super shotgun.
Note that you don't NEED a ProPad controller. It will work fine on a normal
Jaguar pad. You just don't have shoulder buttons, nor Z/Y/X.

The pad controller goes into port 1. When mouse support is done, it will plug
into port 2. You will need an Atari ST or Amiga mouse with appropriate adapter,
or a PS/2 mouse with appropriate adapter for upcoming mouse support.