Doom - Silent No more - After 31 years!
About a week ago, I decided to work on getting music running for a vanilla Jaguar Doom for the community.  After working on this for a while, I was able to get sound working in the demo loop but with issues.  While this started as a side project to attempt this, when I got close, I shared what I had with Chilly Willy.  Chilly was able to identify issues in the sfx.gas file that caused mixing issues with music playing.  He was kind enough to help me update my existing code base, and also get the correct music in the levels.  One difference to note is Chilly left the intermission music in place, while I removed it and have left it to play the into music (not on a loop).  Late yesterday a bug was discovered that crashed the game at level 10, this is fixed in this version.  Also note, this versions sfx.gas is different than the one in JagDoomEx github.  I devised a slightly faster way to mix in the sound effects.  It is with great pleasure, honor, and little me joy to introduce you all to Doom for Atari Jaguar with Music, and Sound.  May God have mercy on you all.  This project started in November of 2024, and has been worked on sporadically by Chilly WIlly, and Tolbat.  This code is free to use, but do not be a dick and give people the credit they deserve if you use it.